requirements.txt
streamlit~=1.38.0
pandas~=2.2.3
numpy
scikit-learn
scipy
matplotlib
databricks-sdk
geopy
requests
